+++ rusqlite/src/trace.rs
@@ -8,7 +8,8 @@ use std::str;
 use std::time::Duration;
 
 use super::ffi;
-use {Error, Result, Connection};
+use {Result, Connection};
+use error::error_from_sqlite_code;
 
 /// Set up the process-wide SQLite error logging callback.
 /// This function is marked unsafe for two reasons:
@@ -42,14 +43,11 @@ pub unsafe fn config_log(callback: Option<fn(c_int, &str)>) -> Result<()> {
         }
     };
 
-    if rc != ffi::SQLITE_OK {
-        return Err(Error {
-            code: rc,
-            message: "sqlite3_config(SQLITE_CONFIG_LOG, ...)".to_string(),
-        });
+    if rc == ffi::SQLITE_OK {
+        Ok(())
+    } else {
+        Err(error_from_sqlite_code(rc, None))
     }
-
-    Ok(())
 }
 
 /// Write a message into the error log established by `config_log`.